Controller working with postman, may need to add application.properties

+++ src/main/java/com/clinicallyinsane/ClinicServer/controller/UserProfileController.java
@@ -41,7 +41,7 @@ public class UserProfileController {
 	ResponseEntity is returned with the User Profile we received
 	 */
 	@GetMapping("/UserProfiles/{id}")
-	public ResponseEntity<UserProfile> getUserProfileById(@PathVariable(value = "id") Long userId, 
+	public ResponseEntity<UserProfile> getUserProfileById(@PathVariable(value = "id") Integer userId, 
 														  @Valid @RequestBody UserProfile userProfileDetails)
 													      throws ResourceNotFoundException{
 		
@@ -66,7 +66,7 @@ public class UserProfileController {
 	
 	 */
 	@PutMapping("/UserProfiles/{id}")
-	public ResponseEntity<UserProfile> updateUserProfile(@PathVariable(value = "id") Long userId, 
+	public ResponseEntity<UserProfile> updateUserProfile(@PathVariable(value = "id") Integer userId, 
 			  											 @Valid @RequestBody UserProfile userProfileDetails)
 			  											 throws ResourceNotFoundException{
 		UserProfile userProfile = userProfileRepository.findById(userId)
@@ -84,7 +84,7 @@ public class UserProfileController {
 	Then, the repository's method delete allows us to pass the found user profile through the method
 	 */
 	@DeleteMapping("/UserProfiles/{id}")
-	public ResponseEntity deleteUserProfile(@PathVariable(value = "id") Long userId) throws ResourceNotFoundException{
+	public ResponseEntity deleteUserProfile(@PathVariable(value = "id") Integer userId) throws ResourceNotFoundException{
 		UserProfile userProfile = userProfileRepository.findById(userId)
 				  .orElseThrow(()-> new ResourceNotFoundException("UserProfile not found with id:" + userId));
 		userProfileRepository.delete(userProfile);
